Better error handling and statefulness.
Connections now have a concept of being "initialized", which means its
connected and authenticated. A connection can become "unitialized" but
it will try to re-initialize itself next time a command is run.

There are probably cases that we could recover from without going
through the drastic step of uninitializing the connection (aka, closing
the socket). But that's not something I'm ready to deal with now,
because it neither seems trivial nor that pressing.